# backend/openapi.json
{"openapi":"3.1.0","info":{"title":"PokéTab API","description":"Backend API for PokéTab - Pokémon Scanning and Collection Platform","version":"1.0.0"},"paths":{"/api/v1/auth/register":{"post":{"tags":["Authentication"],"summary":"Register","description":"Register a new user","operationId":"register_api_v1_auth_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/login":{"post":{"tags":["Authentication"],"summary":"Login","description":"Login user and return JWT tokens","operationId":"login_api_v1_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserLogin"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Token"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/me":{"get":{"tags":["Authentication"],"summary":"Get Current User Info","description":"Get current user information","operationId":"get_current_user_info_api_v1_auth_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/pokemon/scan":{"post":{"tags":["Pokemon"],"summary":"Scan Pokemon","description":"Scan an image to identify a Pokémon\n\nSteps:\n1. Receive image from user\n2. Preprocess image with OpenCV (resize, enhance, denoise)\n3. Use Gemini Vision to identify Pokémon name \n4. Fetch detailed data from PokeAPI\n5. Return complete Pokémon information","operationId":"scan_pokemon_api_v1_pokemon_scan_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_scan_pokemon_api_v1_pokemon_scan_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PokemonResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/pokemon/search/{pokemon_name}":{"get":{"tags":["Pokemon"],"summary":"Search Pokemon","description":"Search for a Pokémon by name","operationId":"search_pokemon_api_v1_pokemon_search__pokemon_name__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"pokemon_name","in":"path","required":true,"schema":{"type":"string","title":"Pokemon Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PokemonResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/collection/":{"get":{"tags":["Collection"],"summary":"Get Collection","description":"Get user's Pokémon collection","operationId":"get_collection_api_v1_collection__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/CollectionResponse"},"type":"array","title":"Response Get Collection Api V1 Collection  Get"}}}}},"security":[{"HTTPBearer":[]}]},"post":{"tags":["Collection"],"summary":"Add To Collection","description":"Add a Pokémon to user's collection","operationId":"add_to_collection_api_v1_collection__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CollectionAddRequest"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CollectionResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/collection/{collection_id}":{"delete":{"tags":["Collection"],"summary":"Remove From Collection","description":"Remove a Pokémon from user's collection","operationId":"remove_from_collection_api_v1_collection__collection_id__delete","security":[{"HTTPBearer":[]}],"parameters":[{"name":"collection_id","in":"path","required":true,"schema":{"type":"integer","title":"Collection Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/collection/count":{"get":{"tags":["Collection"],"summary":"Get Collection Count","description":"Get count of Pokémon in user's collection","operationId":"get_collection_count_api_v1_collection_count_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]}},"/":{"get":{"summary":"Root","description":"Root endpoint","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"summary":"Health Check","description":"Health check endpoint","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"Body_scan_pokemon_api_v1_pokemon_scan_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"}},"type":"object","required":["file"],"title":"Body_scan_pokemon_api_v1_pokemon_scan_post"},"CollectionAddRequest":{"properties":{"pokemon_name":{"type":"string","title":"Pokemon Name"},"pokemon_id":{"type":"integer","title":"Pokemon Id"},"pokemon_data":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Pokemon Data"}},"type":"object","required":["pokemon_name","pokemon_id"],"title":"CollectionAddRequest","description":"Schema for adding Pokémon to collection"},"CollectionResponse":{"properties":{"id":{"type":"integer","title":"Id"},"user_id":{"type":"integer","title":"User Id"},"pokemon_name":{"type":"string","title":"Pokemon Name"},"pokemon_id":{"type":"integer","title":"Pokemon Id"},"pokemon_data":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Pokemon Data"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","user_id","pokemon_name","pokemon_id","created_at"],"title":"CollectionResponse","description":"Schema for collection item response"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"NamedAPIResource":{"properties":{"name":{"type":"string","title":"Name"},"url":{"type":"string","title":"Url"}},"type":"object","required":["name","url"],"title":"NamedAPIResource","description":"Schema for named API resource"},"PokemonAbility":{"properties":{"ability":{"$ref":"#/components/schemas/NamedAPIResource"},"is_hidden":{"type":"boolean","title":"Is Hidden"},"slot":{"type":"integer","title":"Slot"}},"type":"object","required":["ability","is_hidden","slot"],"title":"PokemonAbility","description":"Schema for Pokémon ability"},"PokemonResponse":{"properties":{"id":{"type":"integer","title":"Id"},"name":{"type":"string","title":"Name"},"height":{"type":"integer","title":"Height"},"weight":{"type":"integer","title":"Weight"},"types":{"items":{"$ref":"#/components/schemas/PokemonType"},"type":"array","title":"Types"},"stats":{"items":{"$ref":"#/components/schemas/PokemonStat"},"type":"array","title":"Stats"},"abilities":{"items":{"$ref":"#/components/schemas/PokemonAbility"},"type":"array","title":"Abilities"},"sprites":{"$ref":"#/components/schemas/PokemonSprites"},"species_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Species Url"}},"type":"object","required":["id","name","height","weight","types","stats","abilities","sprites"],"title":"PokemonResponse","description":"Schema for Pokémon information response"},"PokemonSprites":{"properties":{"front_default":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Front Default"},"front_shiny":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Front Shiny"},"other":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Other"}},"type":"object","title":"PokemonSprites","description":"Schema for Pokémon sprites"},"PokemonStat":{"properties":{"base_stat":{"type":"integer","title":"Base Stat"},"effort":{"type":"integer","title":"Effort"},"stat":{"$ref":"#/components/schemas/NamedAPIResource"}},"type":"object","required":["base_stat","effort","stat"],"title":"PokemonStat","description":"Schema for Pokémon stat"},"PokemonType":{"properties":{"slot":{"type":"integer","title":"Slot"},"type":{"$ref":"#/components/schemas/NamedAPIResource"}},"type":"object","required":["slot","type"],"title":"PokemonType","description":"Schema for Pokémon type"},"Token":{"properties":{"access_token":{"type":"string","title":"Access Token"},"refresh_token":{"type":"string","title":"Refresh Token"},"token_type":{"type":"string","title":"Token Type","default":"bearer"}},"type":"object","required":["access_token","refresh_token"],"title":"Token","description":"Schema for JWT token response"},"UserCreate":{"properties":{"username":{"type":"string","maxLength":50,"minLength":3,"title":"Username"},"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","maxLength":100,"minLength":8,"title":"Password"}},"type":"object","required":["username","email","password"],"title":"UserCreate","description":"Schema for user registration"},"UserLogin":{"properties":{"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"}},"type":"object","required":["username","password"],"title":"UserLogin","description":"Schema for user login"},"UserResponse":{"properties":{"id":{"type":"integer","title":"Id"},"username":{"type":"string","title":"Username"},"email":{"type":"string","title":"Email"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","username","email","created_at"],"title":"UserResponse","description":"Schema for user response"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}